feat(v2.1.0): NodeMeta separation, f32 embeddings, reject Binary Quantization
BUG A — Split Node into NodeMeta (~100 bytes) + PoincareVector (embedding)
stored in separate RocksDB column families (CF_NODES + CF_EMBEDDINGS).
BFS traversal reads only NodeMeta per hop (10-25x speedup on large graphs).
update_energy() and update_hausdorff() no longer deserialize embeddings.
Node implements Deref<Target=NodeMeta> for backward compatibility.

ITEM C — Migrate PoincareVector.coords from Vec<f64> to Vec<f32> (50% memory).
Distance kernel (poincare_sums) promotes to f64 internally for numerical
stability near the Poincaré boundary. Proto wire format unchanged (f64).
Sleep cycle and L-System math remain in f64 — promote/narrow at boundaries.

ITEM F — Binary Quantization formally rejected. sign(x) destroys magnitude
which encodes hierarchical position in the Poincaré ball. Decision documented
in lib.rs, risco_hiperbolico.md, and CLAUDE.md. Never implement.

+++ crates/nietzsche-graph/src/transaction.rs
@@ -283,7 +283,7 @@ mod tests {
     fn node(x: f64) -> Node {
         Node::new(
             Uuid::new_v4(),
-            PoincareVector::new(vec![x, 0.0]),
+            PoincareVector::new(vec![x as f32, 0.0]),
             serde_json::json!({}),
         )
     }
